requirements.txt
fastapi~=0.115.8
pymongo~=4.11.2
pydantic~=2.10.6
python-dotenv~=1.0.1
starlette~=0.45.3
mangum~=0.19.0
uvicorn~=0.34.0
passlib~=1.7.4
python-multipart